New types and a connection capability tweak for Broccoli.
This is work to resolve ticket 606, which basically resolves around
Broccoli's ability to receive connection records from Bro.  This
mainly did not work because of lacking type support in Broccoli,
because even though the connection record *value* that Bro sends
may not actually contain most of the optional entries, the underlying
*type* does get sent across in its entirety.  Given all the stuff now
living in connection records, this type recursively contains pretty
much anything Bro has to offer, and required adding enum, file, and
vector types to Broccoli.  These are not currently exposed in the
Broccoli API in any way.

A deeper problem arises when record field attributes in a type contain
expressions, because in order to parse these Broccoli would in principle
need to be able to interpret the entire hierarchy of expressions. Since
that makes no sense, Robin and I agreed to add a new capability that
Broccoli sends at handshake time, which basically identifies the endpoint
as using Broccoli. Bro will simply no longer send the attribute
expressions to such peers.

On the Broccoli side, the new capability is defined as follows ...

  #define BRO_CAP_BROCCOLI_PEER      16

... and gets sent automatically.