Use @rpath in broccoli.dylib's install_name on OS X.
This makes it easier for the library to be relocatable and so more
amenable to other Apple-isms (e.g. frameworks or app bundles).  The cost
may be placing more burden on the "user" of the library (e.g. by making
executables with an RPATH or setting DYLD_LIBRARY_PATH to help the
loader find the library).

(Also make a CMake policy warning go away, but could have done that
without changing install_name).