Add a new option "-r" to the broccoli-v6addrs test program
Sometimes broccoli-v6addrs fails to connect to Bro when run from a test
script (presumably because Bro takes a while to start up).  Added a
new option "-r" to retry upon failure to connect to Bro, which avoids
the race condition and eliminates any need to have a "sleep" between
starting Bro and starting broccoli-v6addrs.